update initialize sdk
