+++ .env
@@ -1,2 +1,19 @@
-PORT=5008
-NODE_ENV='tests'
+URL2 = http://localhost:3001/graphql
+URL = https://hxb-graph.hexabase.com/graphql
+TOKEN2 = eyJhbGciOiJSUzUxMiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2NDk0MTQ0NDQsImlhdCI6MTY0OTA2ODg0NCwic3ViIjoiNjIwZjU3NWJkYjAwMGNiYTk4YjY1ZDczIiwidW4iOiJ1c2VybmFtZSB0ZXN0In0.05Vv6mEBwTb2HZuQmmaaWUF1SksorJC2Xj4NbFC45qeTha13kp7ldXhMPDebWxUI1f_l1-OO3XnTUnqrtuajVhYUcrBIO8jbX19H1un9BWqEiXrtqxj_Rc9eCL79SAmbXYu1ZOHbmGQDRKPWMgN9jkY22W-2XjTWiRhVXa0b6Rr5QBJXlQdI4uHMMNeFTphrHMERV2h6Y309WOxzuj9RBkzTSj6ODNh9LKug1FPblaLou6HW-oIrSPvudj_2LOk3PQZicjlGaqhLQ3oDRNUFZ9V4svG9lC9rif4hOn1HjauT7jVSeOU8P6Iuwp5qGYq4SugpwBrGz5HACZ6MWb0gfg
+TOKEN = eyJhbGciOiJSUzUxMiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2NDk0NjkwNzAsImlhdCI6MTY0OTEyMzQ3MCwic3ViIjoiNjIwZjU3NWJkYjAwMGNiYTk4YjY1ZDczIiwidW4iOiJ1c2VybmFtZSB0ZXN0In0.T7jtCm8jKO7A5L3ZENKnDO3DyYAGjOj7gTFBb4J1Sz01shAWzMmrDgmeSCpWiHSBu3cIjIh-snoL6j2LS1SSsKm9gWhz13oTrEIGovj0O_9M_Y7p7Rav6tbbsz9ckuoEczpRTQeMVvwYsvfifj6SCrR8BxnWNjI4uqckDSF-erKpOWpaErqWyXbgo2TcNQGdck3fj1578WwHEsAsM89ZGmnaFIbdICcxv2s8ZXEFrzCTLDtWV9OMKNqWHi934Dfy7aJlOothqyQBZvWH5qx-eN_xadWM7VMDXzbhCQLB5e1yJD_8lJGHbpLhgkI5h0FUS4Mui1sBDuceY3FKiJmH4A
+WORKSPACEID2 = 620dd1c861985350c2d26e57 
+WORKSPACEID = 624bc3cfa72c942744ee6635
+APPLICATIONID = 624ac2f3cbb42c82793c10e9
+DATASTOREID = 624bff496a315d06bfa34187
+ITEMID = 624d1aca14c1a26dccf4c6cc
+ITEMID1 = 624bff497cd0bfa682c8249b
+ITEMID2 = 624d1b5cf6dca8632435d63c
+FIELDID = 624bff497cd0bfa682c8249b
+ACTIONID = 624bff497cd0bfa682c824a7
+WORKSPACEIDSET = 624bff497cd0bfa682c824a7
+
+TASKID = taskId
+CONFIRMATIONID = lfzrrcv7ajjlm00nzg20rf7rdkleicwn5iga9uezu0e0ajhux4h5hufjuqs8vcqj1lvx4onqwncibr4deg6ouso3zhh0yi3csprl
+EMAIL = nguyenpd@b-eee.com
+PASSWORD = 123456
