+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hxb-sdk",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "main": "dist/index.js",
   "license": "MIT",
   "scripts": {
