+++ src/lib/packages/datastore/datastore.test.ts
@@ -1,4 +1,5 @@
 import Datastore from '.';
+import Auth from '../auth';
 import AuthMw from '../middlware/auth';
 require('dotenv').config();
 /**
@@ -17,8 +18,8 @@ const password = process.env.PASSWORD || '';
 beforeAll( async () => {
   if (email && password) {
     console.log('[email, password]: ', email, password);
-    const authMw = new AuthMw(url);
-    const {token, error} = await authMw.loginAsync({email, password});
+    const auth = new Auth(url);
+    const {token, error} = await auth.loginAsync({email, password});
     if (token) {
       return tokenDs = token;
     } else {
