+++ src/lib/packages/workspace/workspace.test.ts
@@ -1,4 +1,5 @@
 import Workspace from '.';
+import Auth from '../auth';
 import AuthMw from '../middlware/auth';
 require('dotenv').config();
 /**
@@ -26,8 +27,8 @@ const setCurrentWorkSpaceInput = {
 /** run first testing  */
 beforeAll( async () => {
   if (email && password) {
-    const authMw = new AuthMw(url);
-    const {token, error} = await authMw.loginAsync({email, password});
+    const auth = new Auth(url);
+    const {token, error} = await auth.loginAsync({email, password});
     if (token) {
       return tokenWs = token;
     } else {
